==> packages/kptn/kptn-0.1.4.tar.gz/kptn-0.1.4/example/nibrs/tasks/impute_items/part2_nonperson_victims/10_MICE_Item_Missing_Imputation_arr_off_match_v3-non-person.Rmd <==
---
title: "NIBRS_Imputation_item_Missing"
author: "Amang Sukasih / Edited by: Philip Lee"
date: "10/20/2020"
output:
  html_document: default
  pdf_document:  default
---

```{r program10}
options(max.print=100000)
options(tibble.print_max = 1000)
#set.seed(414425)
#Make set.seed = NA to produce consistent results on rerun
set.seed=NA

library(dplyr)
library(dbplyr)
library(tidyverse)
library(haven)
library(mice)
library(miceadds)
library(VIM)
library(naniar)
library(visdat)
library(ggplot2)
library(writexl)
library(foreign)
library(DT)
library(data.table)

options(knit.duplicate.label = "allow")

#input_state <- Sys.getenv("INPUT_STATE")
if (input_state==""){
  simpleError("Set input_state as enviroment variable")
}
input_state <- Sys.getenv("INPUT_STATE")


NIBRS <- logical_edits_frame %>%
  filter(der_data_subset %in% c(2)) %>%
  #Drop relationship_id since there may be multiple victims
  select(-relationship_code)


#See the data to double check
NIBRS %>% checkfunction(
    der_data_subset, der_arrestee_offender_same, der_incident_cleared, der_victim_person, der_victim_non_person, der_victim_person_unknown_off
  )

#Need to drop any victim variables before doing the MICE
log_debug("Checking dimensions of NIBRS")
log_dim(NIBRS)

#Using Amang's code:

NIBRS$sex_code_victim2 <- as.factor(NIBRS$sex_code_victim2)
NIBRS$race_code_victim2 <- as.factor(NIBRS$race_code_victim2)
# NIBRS$relationship_code <- as.factor(NIBRS$relationship_code)
NIBRS$victim_type_code <- as.factor(NIBRS$victim_type_code)
NIBRS$sex_code_offender2 <- as.factor(NIBRS$sex_code_offender2)
NIBRS$race_code_offender2	<- as.factor(NIBRS$race_code_offender2)
NIBRS$sex_code_arrestee2 <- as.factor(NIBRS$sex_code_arrestee2)
NIBRS$race_code_arrestee2	<- as.factor(NIBRS$race_code_arrestee2)
NIBRS$arrest_type_code2 <- as.factor(NIBRS$arrest_type_code2)

NIBRS$der_arrestee_weapon2 <- as.factor(NIBRS$der_arrestee_weapon2)
NIBRS$cleared_except_code	<- as.factor(NIBRS$cleared_except_code)
# NIBRS$crime_against2 <- as.factor(NIBRS$crime_against2)
NIBRS$resident_code2 <- as.factor(NIBRS$resident_code2)
NIBRS$suburban_area_flag2 <- as.factor(NIBRS$suburban_area_flag2)
NIBRS$population_group_id <- as.factor(NIBRS$population_group_id)
# NIBRS$resident_status_code_victim2 <- as.factor(NIBRS$resident_status_code_victim2)
NIBRS$der_bias <- as.factor(NIBRS$der_bias)
NIBRS$der_weapon2 <- as.factor(NIBRS$der_weapon2)
NIBRS$cargo_theft_flag2 <- as.factor(NIBRS$cargo_theft_flag2)
NIBRS$method_entry_code2 <- as.factor(NIBRS$method_entry_code2)
NIBRS$der_arrestee_offender_same <- as.factor(NIBRS$der_arrestee_offender_same)
#NIBRS$offense_code2 <-as.factor(NIBRS$offense_code2)

# For a given arrestee-offender, need to deduplicate due to multiple offense codes (1 victim with more than 1 offense codes), or multiple victims
# NIBRS1adedup <- NIBRS %>%
#   group_by(incident_id, arrestee_seq_num, offender_seq_num) %>%
#   arrange( incident_id, arrestee_seq_num, offender_seq_num, desc(offense_code2)) %>%
#   mutate(raw_first_row = row_number() == 1) %>%
#   filter(raw_first_row == TRUE) %>%
#   ungroup()

#Create the continuous summary counts variables for offense indicator used for MICE

log_debug("Checking dimensions before createoffenserecodeindicator")
log_dim(NIBRS)
NIBRS <- createoffenserecodeindicator(data=NIBRS, type="offender", deduptype="victim")
log_debug("Checking dimensions after createoffenserecodeindicator")
log_dim(NIBRS)

#Declare the factors
NIBRS$der_off_assault_hom_offender <- as.factor(NIBRS$der_off_assault_hom_offender)
NIBRS$der_off_kidnapping_ht_offender <- as.factor(NIBRS$der_off_kidnapping_ht_offender)
NIBRS$der_off_sex_offender <- as.factor(NIBRS$der_off_sex_offender)
NIBRS$der_off_arson_bur_vand_offender <- as.factor(NIBRS$der_off_arson_bur_vand_offender)
NIBRS$der_off_mvt_rob_offender <- as.factor(NIBRS$der_off_mvt_rob_offender)
NIBRS$der_off_other_property_offender <- as.factor(NIBRS$der_off_other_property_offender)
NIBRS$der_off_other_society_offender <- as.factor(NIBRS$der_off_other_society_offender)

NIBRS1adedup <- NIBRS %>%
  group_by(incident_id, arrestee_seq_num, offender_seq_num) %>%
  arrange( incident_id, arrestee_seq_num, offender_seq_num) %>%
  mutate(raw_first_row = row_number() == 1) %>%
  filter(raw_first_row == TRUE) %>%
  ungroup()



log_dim(NIBRS1adedup)
glimpse(NIBRS1adedup)
# after imputation we will merge back imputed values to NIBRS1a

NIBRS <- NIBRS1adedup[,which(names(NIBRS1adedup) %in% c(
"data_year",
"incident_id",
"offender_id",
"arrestee_id2",

"offender_seq_num",
"arrestee_seq_num",

"sex_code_offender2",
"race_code_offender2",
"age_num_offender2",

"sex_code_arrestee2",
"race_code_arrestee2",
"age_num_arrestee2",

#"relationship_code",

"arrest_type_code2",
"der_arrestee_weapon2",
"cleared_except_code",
#"offense_code2",
"crime_against2",
"resident_code2",
"suburban_area_flag2",
"population_group_id",
#"resident_status_code_victim2",
"der_bias",
"der_weapon2",
"cargo_theft_flag2",
"method_entry_code2",
"der_incident_cleared",
"der_arrestee_offender_same",

#list of new recoded offense variables
paste0(listofoffensesrecoded,"_offender")

))]

log_debug("Checking dimensions of NIBRS after merging back to NIBRS")
log_dim(NIBRS)
summary(NIBRS)


#md2 <- as.data.frame(md.pattern(NIBRS))
md2 <- as.data.frame(md.pattern(NIBRS)) %>% mutate(id = rownames(.))

#Check if imputation is needed
raw_numbercheck <- md2[1,] %>%
  mutate(new_id = str_match(.[,"id"], "X?(\\d+)")[2]) %>%  #Get the number portion
  select(new_id) %>%
  as.double() #Get the number of cells that are non-empty
raw_numberrows <- nrow(NIBRS) %>% as.double() #Get the number of total records

#If true then output dataset and no imputation needed
if(raw_numbercheck == raw_numberrows |

   #Check if any imputation is needed for offender and arrestee variables
   (any(is.na(NIBRS$sex_code_offender2)) == FALSE &
    any(is.na(NIBRS$race_code_offender2))== FALSE &
    any(is.na(NIBRS$age_num_offender2))  == FALSE &
    any(is.na(NIBRS$sex_code_arrestee2)) == FALSE &
    any(is.na(NIBRS$race_code_arrestee2))  == FALSE &
    any(is.na(NIBRS$age_num_arrestee2))  == FALSE
    )) {

NIBRS %>%
    select(incident_id, offender_seq_num, sex_code_offender2, race_code_offender2, age_num_offender2,
                        arrestee_seq_num, sex_code_arrestee2, race_code_arrestee2, age_num_arrestee2
           ) %>%
    write_csv_logging(paste0(mainpath, "10_", input_state, "_imputed_arrestee_offender_match.csv.gz"), na="" )

NIBRS %>%
    write_csv_logging(paste0(mainpath, "10_", input_state, "_mice_arrestee_offender_match.csv.gz"), na="" )  



}else{
  
#Create row indicator variable for merging
NIBRS <- NIBRS %>%
  mutate(
    tbd_row_number = row_number()
  )

#Declare the arrestee variables
CONST_ARRESTEE_VARS <- c("arrestee_seq_num", "arrestee_id2", 
                         "sex_code_arrestee2", "age_num_arrestee2", "race_code_arrestee2")

#Need to overwrite NIBRS and drop the arrestee variables
NIBRS_arrestee_vars <- NIBRS %>%
  select(tbd_row_number, data_year, incident_id, 
         !!!(CONST_ARRESTEE_VARS %>% rlang:::parse_exprs()))

#Drop the arrestee variables from NIBRS
log_debug("Checking dimensions of NIBRS before dropping arrestee variables")
log_dim(NIBRS)

log_debug("Checking dimensions of NIBRS after dropping arrestee variables")
NIBRS <- NIBRS %>%
  select(!!!(paste0("-", CONST_ARRESTEE_VARS) %>% rlang:::parse_exprs()))

log_dim(NIBRS)


write_xlsx_logging(md2,paste0(miscPathout ,"10_mdpattern2_", input_state, ".xlsx"))

init1 = mice(NIBRS, maxit=0)
meth1 = init1$method
pred1 = init1$predictorMatrix

meth2 <- meth1

# meth2[c("sex_code_offender2","race_code_offender2","age_num_offender2",
#         "sex_code_arrestee2","race_code_arrestee2","age_num_arrestee2")] <- c(rep("pmm",6))

#Need to drop arrestee
meth2[c("sex_code_offender2","race_code_offender2","age_num_offender2")] <- c(rep("pmm",3))


## MODEL 1 (only offender demographic vars were imputed, and summary of victim demog vars were used as predictors)

# incpreds <- c("sex_code_offender2","race_code_offender2","age_num_offender2",
#               "sex_code_arrestee2","race_code_arrestee2","age_num_arrestee2")

incpreds <- c("sex_code_offender2","race_code_offender2","age_num_offender2")

predQuick1 <- quickpred(NIBRS, minpuc = 0.25, include = incpreds)
# predQuick1[,c("data_year","incident_id","offender_id","arrestee_id2",
#               "offender_seq_num","arrestee_seq_num",
#               "der_incident_cleared","der_arrestee_offender_same")] <- 0

predQuick1[,c("data_year","incident_id","offender_id",
              "offender_seq_num", 
              "der_incident_cleared","der_arrestee_offender_same", "tbd_row_number")] <- 0

write_xlsx_logging(as.data.frame(predQuick1), paste0(miscPathout, "10_predQuick1_", input_state, ".xlsx"))
write_xlsx_logging(as.data.frame(meth1),      paste0(miscPathout ,"10_meth1", input_state, ".xlsx"))
write_xlsx_logging(as.data.frame(meth2),      paste0(miscPathout ,"10_meth2", input_state, ".xlsx"))


#imp1 = mice(NIBRS, pred=predQuick1, meth=meth1, m=1, seed=414425, maxit=15)
imp1 = mice(NIBRS, pred=predQuick1, meth="pmm", m=1, seed=414425, maxit=15)

#plot(imp1,c("sex_code_offender2","race_code_offender2","age_num_offender2"))
NIBRS1 <- complete(imp1,1)
#NIBRS2 <- complete(imp1,2)
#NIBRS3 <- complete(imp1,3)
#NIBRS4 <- complete(imp1,4)
#NIBRS5 <- complete(imp1,5)

#20220720 - Code to fix arrestee's age and any other missing arrestee's demographic
#if(any(is.na(NIBRS1$age_num_arrestee2)) | any(is.na(NIBRS1$race_code_arrestee2)) | any(is.na(NIBRS1$sex_code_arrestee2)) ){
#  imp1_2 = mice(NIBRS1, pred=predQuick1, meth="pmm", m=1, seed=414425, maxit=15, threshold = 1.1)
#  NIBRS1 <- complete(imp1_2,1)
#}

#Using NIBRS1, need to merge on the arrestee variables
log_debug("Checking dimensions of NIBRS1 after MICE")
log_dim(NIBRS1)

NIBRS1 <- NIBRS1 %>%
  left_join(NIBRS_arrestee_vars, by=c("data_year", "incident_id", "tbd_row_number")) %>%
  select(-tbd_row_number)

log_debug("Checking dimensions of NIBRS1 after merging on the arrestee variables")
log_dim(NIBRS1)


check0 <- NIBRS[1:100,]
write_xlsx_logging(check0,paste0(miscPathout, "10_check0_", input_state, ".xlsx"))

check1 <- NIBRS1[1:100,]
write_xlsx_logging(check1, paste0(miscPathout, "10_check1_", input_state, ".xlsx"))

NIBRSimp <- as.data.frame(NIBRS1) %>%
  mutate(
    #Fix any missing
    ########################Offender#######################################
    sex_code_offender2 = fcase(
      !is.na(sex_code_offender2), sex_code_offender2, 
      is.na(sex_code_offender2),  sex_code_arrestee2 
    ),
    
    race_code_offender2 = fcase(
      !is.na(race_code_offender2), race_code_offender2, 
      is.na(race_code_offender2),  race_code_arrestee2 
    ),    
    
    age_num_offender2 = fcase(
      !is.na(age_num_offender2), age_num_offender2, 
      is.na(age_num_offender2),  age_num_arrestee2 
    ),    
    
    ########################Arrestee#######################################
    sex_code_arrestee2 = fcase(
      !is.na(sex_code_arrestee2), sex_code_arrestee2, 
      is.na(sex_code_arrestee2),  sex_code_offender2 
    ),
    
    race_code_arrestee2 = fcase(
      !is.na(race_code_arrestee2), race_code_arrestee2, 
      is.na(race_code_arrestee2),  race_code_offender2 
    ),
    
    age_num_arrestee2 = fcase(
      !is.na(age_num_arrestee2), age_num_arrestee2, 
      is.na(age_num_arrestee2),  age_num_offender2 
    )
  )

# Becaue all offender=arrestee (correlation of both demographic vars = 1), then MICE will impute only arrestee's demog vars,
# then edit demog vars of offenders after imputation

# Edit offender or arrestee demog var for same person
# if(any(is.na(NIBRSimp$sex_code_arrestee2))==FALSE & any(is.na(NIBRSimp$sex_code_offender2))==TRUE){
#    NIBRSimp$sex_code_offender2 <- NIBRSimp$sex_code_arrestee2}
# if(any(is.na(NIBRSimp$sex_code_arrestee2))==TRUE & any(is.na(NIBRSimp$sex_code_offender2))==FALSE){
#    NIBRSimp$sex_code_arrestee2 <- NIBRSimp$sex_code_offender2}
# 
# if(any(is.na(NIBRSimp$race_code_arrestee2))==FALSE & any(is.na(NIBRSimp$race_code_offender2))==TRUE){
#    NIBRSimp$race_code_offender2 <- NIBRSimp$race_code_arrestee2}
# if(any(is.na(NIBRSimp$race_code_arrestee2))==TRUE & any(is.na(NIBRSimp$race_code_offender2))==FALSE){
#    NIBRSimp$race_code_arrestee2 <- NIBRSimp$race_code_offender2}
# 
# if(any(is.na(NIBRSimp$age_num_arrestee2))==FALSE & any(is.na(NIBRSimp$age_num_offender2))==TRUE){
#    NIBRSimp$age_num_offender2 <- NIBRSimp$age_num_arrestee2}
# if(any(is.na(NIBRSimp$age_num_arrestee2))==TRUE & any(is.na(NIBRSimp$age_num_offender2))==FALSE){
#    NIBRSimp$age_num_arrestee2 <- NIBRSimp$age_num_offender2}

checkfunction(NIBRSimp,sex_code_offender2,sex_code_arrestee2)
checkfunction(NIBRSimp,race_code_offender2 ,race_code_arrestee2 )
checkfunction(NIBRSimp,age_num_offender2 ,age_num_arrestee2 )


#Write out dataset to the share
NIBRSimp %>%
    select(incident_id, offender_seq_num, sex_code_offender2, race_code_offender2, age_num_offender2,
                        arrestee_seq_num, sex_code_arrestee2, race_code_arrestee2, age_num_arrestee2
           ) %>%
    write_csv_logging(paste0(mainpath, "10_", input_state, "_imputed_arrestee_offender_match.csv.gz"), na="" )

NIBRSimp %>%
    write_csv_logging(paste0(mainpath, "10_", input_state, "_mice_arrestee_offender_match.csv.gz"), na="" ) 

}


```
